Workflow is not a post-processing unit, move it up
