+++ lib/post_processing.js
@@ -14,7 +14,7 @@ var path = require('path');
 var execAsync = require('../utils/exec').async;
 
 // General purpose helpers controlling the post-processing workflow.
-var _workflow = require('./units/workflow');
+var _workflow = require('./workflow');
 
 // Available post-processing units, as an ordered stack.
 // NOTE: could probably be implemented as a middleware pipeline, with the last
