+++ lib/workflow.js
@@ -2,14 +2,14 @@
  * **Provides junctures for the post-processing workflow**. Think hooks and
  * normalized operations.
  *
- * @module post-processing/workflow
+ * @module workflow
  */
 
 var fs = require('fs');
 var path = require('path');
 var tmp = require('tmp');
 
-var execAsync = require('../../utils/exec').async;
+var execAsync = require('../utils/exec').async;
